Fixed what I broke
Readthedocs will now build again since the `mkdocs.yml` isn't
referencing a non-existent file anymore. Not merging this means the 1.0
section of the guide will be "frozen" in place, as it currently is,
which'd allow for putting off the release of 1.0 a little longer.

Also updated the JSON Explanations to match what's now generated/used in
1.0
The `Db` section is commented out for the sake of having to change less
in the future.